feat: Phases 4+5+6 — CPU backend, CLI, GGUF I/O, all tests pass
- CPU backend with rayon parallelism and SIMD support
- Full CLI with 7 subcommands: compress, verify, bench, calibrate, audit, info, daemon
- GGUF I/O framework (parser, writer, types)
- 42 unit tests + 8 doc tests — all green
- Clean build with zero warnings

+++ rust/crates/turboquant-core/src/bitpack.rs
@@ -178,7 +178,7 @@ mod tests {
 
     #[test]
     fn roundtrip_slice_4096() {
-        let values: Vec<u8> = (0..4096u8).map(|x| x % 8).collect();
+        let values: Vec<u8> = (0..4096u16).map(|x| (x % 8) as u8).collect();
         let packed = pack_3bit_slice(&values).unwrap();
         let unpacked = unpack_3bit_slice(&packed, 4096).unwrap();
         assert_eq!(values, unpacked);
